+++ packages/twenty-server/src/core/messaging/timeline-messaging.resolver.ts
@@ -1,68 +1,11 @@
-import { Args, Query, Field, Resolver, ObjectType, Int } from '@nestjs/graphql';
+import { Args, Query, Resolver, Int } from '@nestjs/graphql';
 import { UseGuards } from '@nestjs/common';
-
-import { Entity } from 'typeorm';
 
 import { JwtAuthGuard } from 'src/guards/jwt.auth.guard';
 import { Workspace } from 'src/core/workspace/workspace.entity';
 import { AuthWorkspace } from 'src/decorators/auth-workspace.decorator';
 import { TimelineMessagingService } from 'src/core/messaging/timeline-messaging.service';
-
-@Entity({ name: 'timelineThreadParticipant', schema: 'core' })
-@ObjectType('TimelineThreadParticipant')
-class TimelineThreadParticipant {
-  @Field({ nullable: true })
-  personId: string;
-
-  @Field({ nullable: true })
-  workspaceMemberId: string;
-
-  @Field()
-  firstName: string;
-
-  @Field()
-  lastName: string;
-
-  @Field()
-  displayName: string;
-
-  @Field()
-  avatarUrl: string;
-
-  @Field()
-  handle: string;
-}
-
-@Entity({ name: 'timelineThread', schema: 'core' })
-@ObjectType('TimelineThread')
-export class TimelineThread {
-  @Field()
-  id: string;
-
-  @Field()
-  read: boolean;
-
-  @Field()
-  firstParticipant: TimelineThreadParticipant;
-
-  @Field(() => [TimelineThreadParticipant])
-  lastTwoParticipants: TimelineThreadParticipant[];
-
-  @Field()
-  lastMessageReceivedAt: Date;
-
-  @Field()
-  lastMessageBody: string;
-
-  @Field()
-  subject: string;
-
-  @Field()
-  numberOfMessagesInThread: number;
-
-  @Field()
-  participantCount: number;
-}
+import { TimelineThread } from 'src/core/messaging/timeline-thread.entity';
 
 @UseGuards(JwtAuthGuard)
 @Resolver(() => [TimelineThread])
